write tests between logic and data layer

+++ app/src/main/java/com/group12/fitnics/persistence/hsqldb/ExercisePersistenceHSQLDB.java
@@ -2,6 +2,10 @@ package com.group12.fitnics.persistence.hsqldb;
 
 import android.util.Log;
 
+import com.group12.fitnics.exceptions.ExerciseNotFoundException;
+import com.group12.fitnics.exceptions.FoodNotFoundException;
+import com.group12.fitnics.exceptions.InvalidExNameException;
+import com.group12.fitnics.exceptions.InvalidExerciseException;
 import com.group12.fitnics.objects.Exercise;
 import com.group12.fitnics.persistence.IExercisePersistence;
 
@@ -12,6 +16,8 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 
 public class ExercisePersistenceHSQLDB implements IExercisePersistence {
@@ -40,22 +46,34 @@ public class ExercisePersistenceHSQLDB implements IExercisePersistence {
 
     @Override
     public List<Exercise> getExerciseList() {
-        final List<Exercise> exercises = new ArrayList<>();
+        final List<Exercise> exerciseList = new ArrayList<>();
         try (final Connection c = connect()) {
             final Statement st = c.createStatement();
             final ResultSet rs = st.executeQuery("SELECT * FROM EXERCISES");
             while (rs.next()) {
                 final Exercise exercise = fromResultSet(rs);
-                exercises.add(exercise);
+                exerciseList.add(exercise);
             }
             rs.close();
             st.close();
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
-        return exercises;
+
+        List<Exercise> exerciseSortedAlpha = new ArrayList<>(exerciseList.size());
+
+        exerciseSortedAlpha.addAll(exerciseList);
+
+        Collections.sort(exerciseSortedAlpha, new Comparator<Exercise>() {
+            @Override
+            public int compare(Exercise exercise1, Exercise exercise2) {
+                return (exercise1.getTitle()).compareToIgnoreCase(exercise2.getTitle());
+            }
+        });
+
+        return exerciseSortedAlpha;
     }
 
     @Override
@@ -69,7 +87,7 @@ public class ExercisePersistenceHSQLDB implements IExercisePersistence {
             }
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
         return null;
@@ -78,7 +96,7 @@ public class ExercisePersistenceHSQLDB implements IExercisePersistence {
     @Override
     public Exercise getExerciseByName(String exerciseName) {
         try (Connection c = connect()) {
-            final PreparedStatement st = c.prepareStatement("SELECT * FROM EXERCISES WHERE name = ?");
+            final PreparedStatement st = c.prepareStatement("SELECT * FROM EXERCISES WHERE LOWER(name) = LOWER(?)");
             st.setString(1, exerciseName);
             final ResultSet rs = st.executeQuery();
             if (rs.next()) {
@@ -86,7 +104,7 @@ public class ExercisePersistenceHSQLDB implements IExercisePersistence {
             }
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
         return null;
@@ -107,7 +125,7 @@ public class ExercisePersistenceHSQLDB implements IExercisePersistence {
             st.close();
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
 
@@ -129,7 +147,7 @@ public class ExercisePersistenceHSQLDB implements IExercisePersistence {
             st.close();
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
 
@@ -137,25 +155,48 @@ public class ExercisePersistenceHSQLDB implements IExercisePersistence {
     }
 
     @Override
-    public void insertExercise(Exercise newExercise) {
+    public int insertExercise(Exercise newExercise) throws InvalidExerciseException {
+        if(newExercise == null)
+            throw new InvalidExerciseException("The exercise is not valid. ");
+        if(newExercise.getTitle().length() > 20)
+            throw new InvalidExNameException("The name should be no more than 20 characters.");
+        if(newExercise.getDescription().length() > 1024)
+            throw new InvalidExNameException("The description should be no more than 1024 characters.");
+        if(newExercise.getLevel().length() > 20)
+            throw new InvalidExNameException("The level should be no more than 20 characters.");
+        if(newExercise.getCategory().length() > 20)
+            throw new InvalidExNameException("The category should be no more than 20 characters.");
+
         try (final Connection c = connect()) {
-            final PreparedStatement st = c.prepareStatement("INSERT INTO EXERCISES VALUES(DEFAULT, ?, ?, ?, ?, ?)");
-            st.setString(1, newExercise.getTitle());
-            st.setString(2, newExercise.getDescription());
-            st.setString(3, newExercise.getCategory());
-            st.setString(4, newExercise.getLevel());
-            st.setInt(5, newExercise.getCaloriesBurn());
+            final PreparedStatement st = c.prepareStatement("INSERT INTO EXERCISES VALUES(?, ?, ?, ?, ?, ?)");
+
+            newExercise.setExerciseID();
+
+            st.setString(1, Integer.toString(newExercise.getExerciseID()));
+            st.setString(2, newExercise.getTitle());
+            st.setString(3, newExercise.getDescription());
+            st.setString(4, newExercise.getCategory());
+            st.setString(5, newExercise.getLevel());
+            st.setInt(6, newExercise.getCaloriesBurn());
             st.executeUpdate();
             st.close();
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
+
+        return newExercise.getExerciseID();
     }
 
     @Override
-    public void updateExercise(int exerciseID, Exercise currentExercise) {
+    public void updateExercise(int exerciseID, Exercise currentExercise) throws InvalidExerciseException, ExerciseNotFoundException {
+        if(currentExercise == null)
+            throw new InvalidExerciseException("The exercise is not valid. ");
+
+        if (getExerciseById(exerciseID) == null)
+            throw new ExerciseNotFoundException("There's no exercise with the exerciseID. ");
+
         try (final Connection c = connect()) {
             final PreparedStatement st = c.prepareStatement(
                     "UPDATE EXERCISES SET name=? description=? category=? level=? calPerMin=? where id = ?");
@@ -169,34 +210,33 @@ public class ExercisePersistenceHSQLDB implements IExercisePersistence {
             st.close();
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
     }
 
     @Override
-    public void deleteExercise(int exerciseID) {
-        try (final Connection c = connect()) {
-            // Delete on cascade
-            final PreparedStatement eLogsSt = c.prepareStatement("DELETE FROM EXERCISELOGS WHERE eid = ?");
-            eLogsSt.setString(1, Integer.toString(exerciseID));
-            eLogsSt.executeUpdate();
-            eLogsSt.close();
+    public void deleteExercise(int exerciseID) throws ExerciseNotFoundException {
+        if (getExerciseById(exerciseID) == null)
+            throw new ExerciseNotFoundException("There's no exercise with the exerciseID. ");
 
+        try (final Connection c = connect()) {
             final PreparedStatement st = c.prepareStatement("DELETE FROM EXERCISES WHERE id = ?");
             st.setString(1, Integer.toString(exerciseID));
             st.executeUpdate();
             st.close();
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
     }
 
     @Override
-    public void deleteExercise(Exercise currentExercise) {
-        int exerciseID = currentExercise.getExerciseID();
-        deleteExercise(exerciseID);
+    public void deleteExercise(Exercise currentExercise) throws ExerciseNotFoundException {
+        if (getExerciseById(currentExercise.getExerciseID()) == null)
+            throw new ExerciseNotFoundException("There's no exercise to delete. ");
+
+        deleteExercise(currentExercise.getExerciseID());
     }
 }
