+++ app/src/main/java/com/group12/fitnics/persistence/hsqldb/UserPersistenceHSQLDB.java
@@ -1,7 +1,8 @@
 package com.group12.fitnics.persistence.hsqldb;
 
-import android.util.Log;
-
+import com.group12.fitnics.exceptions.InvalidUserException;
+import com.group12.fitnics.exceptions.InvalidUsernameException;
+import com.group12.fitnics.exceptions.UserNotFoundException;
 import com.group12.fitnics.objects.User;
 import com.group12.fitnics.persistence.IUserPersistence;
 
@@ -38,9 +39,13 @@ public class UserPersistenceHSQLDB implements IUserPersistence {
         char gender = rs.getString("gender").charAt(0);
         Integer dailyCaloricNeeds = rs.getInt("caloricNeeds");
         Integer goal = rs.getInt("goal");
-//        String unit = rs.getString("unit"); // ?????
+        int[] units = {0, 0};
+        if (rs.getString("wUnit").equals("kg"))
+            units[0] = 1;
+        if (rs.getString("hUnit").equals("ft"))
+            units[1] = 1;
 
-        User user = new User(username, activityLvl, weight, height, gender);
+        User user = new User(username, birthDay, birthMonth, birthYear, activityLvl, weight, height, gender, goal, units);
         user.setUserID(userID);
         return user;
     }
@@ -59,7 +64,7 @@ public class UserPersistenceHSQLDB implements IUserPersistence {
             st.close();
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
         return users;
@@ -76,7 +81,7 @@ public class UserPersistenceHSQLDB implements IUserPersistence {
             }
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
         return null;
@@ -93,34 +98,70 @@ public class UserPersistenceHSQLDB implements IUserPersistence {
             }
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
         return null;
     }
 
     @Override
-    public void insertUser(User currentUser) {
+    public int insertUser(User currentUser) throws InvalidUserException {
+        if (currentUser == null)
+            throw new InvalidUserException("The user is not valid. ");
+
+        // if there exists same username already, do not allow it to be inserted
+        if (getUserByUsername(currentUser.getUsername()) != null)
+            throw new InvalidUsernameException("There exists duplicate username. ");
+
+        if (currentUser.getUsername().length() > 20)
+            throw new InvalidUsernameException("The username is too long, should be no more than 20 characters.");
+
         try (final Connection c = connect()) {
-            final PreparedStatement st = c.prepareStatement("INSERT INTO USERS VALUES(DEFAULT, ?, ?, ?, ?, ?)");
-            st.setString(1, currentUser.getUsername());
-            st.setInt(2, currentUser.getActivityLevel());
-            st.setDouble(3, currentUser.getWeight());
-            st.setDouble(4, currentUser.getHeight());
-            st.setString(5, String.valueOf(currentUser.getGender()));
+            final PreparedStatement st = c.prepareStatement("INSERT INTO USERS VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
+
+            currentUser.setUserID();
+
+            st.setString(1, Integer.toString(currentUser.getUserID()));
+            st.setString(2, currentUser.getUsername());
+            st.setString(3, Integer.toString(currentUser.getBirthDay()));
+            st.setString(4, Integer.toString(currentUser.getBirthMonth()));
+            st.setString(5, Integer.toString(currentUser.getBirthYear()));
+            st.setInt(6, currentUser.getActivityLevel());
+            st.setDouble(7, currentUser.getWeight());
+            st.setDouble(8, currentUser.getHeight());
+            st.setString(9, String.valueOf(currentUser.getGender()));
+            st.setInt(10, currentUser.getDailyCaloricNeeds());
+            st.setInt(11, currentUser.getGoal());
+            String wUnit = "lbs";
+            String hUnit = "cm";
+            if (currentUser.getUnits()[0] == 1)
+                wUnit = "kg";
+            if (currentUser.getUnits()[1] == 1)
+                hUnit = "ft";
+            st.setString(12, wUnit);
+            st.setString(13, hUnit);
+
             st.executeUpdate();
             st.close();
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
+
+        return currentUser.getUserID();
     }
 
     @Override
-    public void updateUser(int userID, User currentUser) {
+    public void updateUser(int userID, User currentUser) throws InvalidUserException, UserNotFoundException {
+        if (currentUser == null)
+            throw new InvalidUserException("The user is not valid. ");
+
+        if (getUserByID(userID) == null)
+            throw new UserNotFoundException("There's no user with the userID. ");
+
         try (final Connection c = connect()) {
-            final PreparedStatement st = c.prepareStatement("UPDATE USERS SET username=?, activityLvl=? weight=? height=? gender=? caloricNeeds=? goal=? WHERE id = ?");
+            final PreparedStatement st = c.prepareStatement("UPDATE USERS SET username=?, actLvl=?, weight=?, height=?, gender=?, caloricNeeds=?, goal=? WHERE id = ?");
             st.setString(1, currentUser.getUsername());
             st.setInt(2, currentUser.getActivityLevel());
             st.setDouble(3, currentUser.getWeight());
@@ -133,21 +174,24 @@ public class UserPersistenceHSQLDB implements IUserPersistence {
             st.close();
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
     }
 
     @Override
-    public void deleteUser(int userID) {
+    public void deleteUser(int userID) throws UserNotFoundException {
+        if (getUserByID(userID) == null)
+            throw new UserNotFoundException("There's no user with the userID. ");
+
         try (final Connection c = connect()) {
-            // Delete on cascade ???
+            // Delete on cascade
             final PreparedStatement foodlogs = c.prepareStatement("DELETE FROM FOODLOGS WHERE uid = ?");
             foodlogs.setString(1, Integer.toString(userID));
             foodlogs.executeUpdate();
             foodlogs.close();
 
-            // Delete on cascade ???
+            // Delete on cascade
             final PreparedStatement excslogs = c.prepareStatement("DELETE FROM EXERCISELOGS WHERE uid = ?");
             excslogs.setString(1, Integer.toString(userID));
             excslogs.executeUpdate();
@@ -159,7 +203,7 @@ public class UserPersistenceHSQLDB implements IUserPersistence {
             st.close();
 
         } catch (final SQLException e) {
-            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+//            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
             e.printStackTrace();
         }
     }
